model_serve/facenet/requirements.txt
opencv-python
scipy
torch
pillow
tensorrt
torch-tensorrt